Improve Pool's graceful shutdown
Previously, there was an indirection of making the master
report to itself and used a delay in a bid to optimistically
allow workers shutdown with no acknowlegments.

This change ensures that each worker has sent a finished
shutting down acknowledgement to master before
breaking out of the event loop, allowing the pool to
shutdown gracefully.