Extract worker's result agent
This change renames `worker_assistant` to `result_agent` which explicitly
describes what the extra Actor is for. The `result_agent` has just one task per
worker, and that is to redirect the results to the master. It was introduced to
avoid blocking the main worker process when processing the blocking task
of executing transactions.